+++ src/Checkout/Plus/PlusPaymentFinalizeController.php
@@ -7,6 +7,7 @@
 
 namespace Swag\PayPal\Checkout\Plus;
 
+use Psr\Log\LoggerInterface;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionEntity;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStateHandler;
 use Shopware\Core\Checkout\Payment\Cart\AsyncPaymentTransactionStruct;
@@ -14,12 +15,14 @@ use Shopware\Core\Checkout\Payment\Cart\PaymentHandler\AsynchronousPaymentHandle
 use Shopware\Core\Checkout\Payment\Exception\CustomerCanceledAsyncPaymentException;
 use Shopware\Core\Checkout\Payment\Exception\InvalidTransactionException;
 use Shopware\Core\Checkout\Payment\Exception\PaymentProcessException;
+use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\MultiFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\NotFilter;
 use Shopware\Core\Framework\Routing\Annotation\RouteScope;
+use Shopware\Core\Framework\Routing\Annotation\Since;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 use Swag\PayPal\Checkout\Payment\PayPalPaymentHandler;
 use Swag\PayPal\SwagPayPal;
@@ -51,16 +54,23 @@ class PlusPaymentFinalizeController extends AbstractController
      */
     private $router;
 
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
     public function __construct(
         EntityRepositoryInterface $orderTransactionRepo,
         AsynchronousPaymentHandlerInterface $paymentHandler,
         OrderTransactionStateHandler $transactionStateHandler,
-        RouterInterface $router
+        RouterInterface $router,
+        LoggerInterface $logger
     ) {
         $this->orderTransactionRepo = $orderTransactionRepo;
         $this->paymentHandler = $paymentHandler;
         $this->transactionStateHandler = $transactionStateHandler;
         $this->router = $router;
+        $this->logger = $logger;
     }
 
     /**
@@ -101,10 +111,10 @@ class PlusPaymentFinalizeController extends AbstractController
                 ]
             )
         );
-        $result = $this->orderTransactionRepo->search($criteria, $salesChannelContext->getContext());
 
+        $context = $salesChannelContext->getContext();
         /** @var OrderTransactionEntity|null $orderTransaction */
-        $orderTransaction = $result->getEntities()->first();
+        $orderTransaction = $this->orderTransactionRepo->search($criteria, $context)->getEntities()->first();
 
         if ($orderTransaction === null) {
             throw new InvalidTransactionException('');
@@ -128,18 +138,77 @@ class PlusPaymentFinalizeController extends AbstractController
         try {
             $this->paymentHandler->finalize($paymentTransactionStruct, $request, $salesChannelContext);
         } catch (PaymentProcessException $paymentProcessException) {
-            $this->transactionStateHandler->fail(
-                $paymentProcessException->getOrderTransactionId(),
-                $salesChannelContext->getContext()
-            );
-            $finishUrl = $this->router->generate('frontend.checkout.finish.page', [
-                'orderId' => $orderId,
-                PayPalPaymentHandler::PAYPAL_PLUS_CHECKOUT_ID => true,
-                'changedPayment' => $changedPayment,
-                'paymentFailed' => true,
-            ]);
+            if ($this->isAtLeastShopware6330()) {
+                $finishUrl = $this->redirectToConfirmPageWorkflow(
+                    $paymentProcessException,
+                    $context,
+                    $orderId
+                );
+            } else {
+                $finishUrl = $this->redirectToFinishPageWorkflow(
+                    $paymentProcessException,
+                    $context,
+                    $orderId,
+                    $changedPayment
+                );
+            }
         }
 
         return new RedirectResponse($finishUrl);
+    }
+
+    private function redirectToConfirmPageWorkflow(
+        PaymentProcessException $paymentProcessException,
+        Context $context,
+        string $orderId
+    ): string {
+        $errorUrl = $this->router->generate('frontend.account.edit-order.page', ['orderId' => $orderId]);
+
+        if ($paymentProcessException instanceof CustomerCanceledAsyncPaymentException) {
+            $this->transactionStateHandler->cancel(
+                $paymentProcessException->getOrderTransactionId(),
+                $context
+            );
+            $urlQuery = \parse_url($errorUrl, PHP_URL_QUERY) ? '&' : '?';
+
+            return \sprintf('%s%serror-code=%s', $errorUrl, $urlQuery, $paymentProcessException->getErrorCode());
+        }
+
+        $transactionId = $paymentProcessException->getOrderTransactionId();
+        $this->logger->error(
+            'An error occurred during finalizing async payment',
+            ['orderTransactionId' => $transactionId, 'exceptionMessage' => $paymentProcessException->getMessage()]
+        );
+        $this->transactionStateHandler->fail(
+            $transactionId,
+            $context
+        );
+        $urlQuery = \parse_url($errorUrl, PHP_URL_QUERY) ? '&' : '?';
+
+        return \sprintf('%s%serror-code=%s', $errorUrl, $urlQuery, $paymentProcessException->getErrorCode());
+    }
+
+    private function redirectToFinishPageWorkflow(
+        PaymentProcessException $paymentProcessException,
+        Context $context,
+        string $orderId,
+        bool $changedPayment
+    ): string {
+        $this->transactionStateHandler->fail(
+            $paymentProcessException->getOrderTransactionId(),
+            $context
+        );
+
+        return $this->router->generate('frontend.checkout.finish.page', [
+            'orderId' => $orderId,
+            PayPalPaymentHandler::PAYPAL_PLUS_CHECKOUT_ID => true,
+            'changedPayment' => $changedPayment,
+            'paymentFailed' => true,
+        ]);
+    }
+
+    private function isAtLeastShopware6330(): bool
+    {
+        return \class_exists(Since::class);
     }
 }