+++ tests/Mock/Payment/AsyncPaymentHandlerMock.php
@@ -9,14 +9,18 @@ namespace Swag\PayPal\Test\Mock\Payment;
 
 use Shopware\Core\Checkout\Payment\Cart\AsyncPaymentTransactionStruct;
 use Shopware\Core\Checkout\Payment\Cart\PaymentHandler\AsynchronousPaymentHandlerInterface;
+use Shopware\Core\Checkout\Payment\Exception\AsyncPaymentFinalizeException;
 use Shopware\Core\Checkout\Payment\Exception\CustomerCanceledAsyncPaymentException;
 use Shopware\Core\Framework\Validation\DataBag\RequestDataBag;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Swag\PayPal\Checkout\Payment\PayPalPaymentHandler;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 
 class AsyncPaymentHandlerMock implements AsynchronousPaymentHandlerInterface
 {
+    public const REQUEST_PARAM_SHOULD_THROW_EXCEPTION = 'throwException';
+
     public function pay(
         AsyncPaymentTransactionStruct $transaction,
         RequestDataBag $dataBag,
@@ -29,11 +33,16 @@ class AsyncPaymentHandlerMock implements AsynchronousPaymentHandlerInterface
         Request $request,
         SalesChannelContext $salesChannelContext
     ): void {
-        if ($request->query->getBoolean('cancel')) {
+        $transactionId = $transaction->getOrderTransaction()->getId();
+        if ($request->query->getBoolean(PayPalPaymentHandler::PAYPAL_REQUEST_PARAMETER_CANCEL)) {
             throw new CustomerCanceledAsyncPaymentException(
-                $transaction->getOrderTransaction()->getId(),
+                $transactionId,
                 'Customer canceled the payment on the PayPal page'
             );
+        }
+
+        if ($request->query->getBoolean(self::REQUEST_PARAM_SHOULD_THROW_EXCEPTION)) {
+            throw new AsyncPaymentFinalizeException($transactionId, 'Test error message');
         }
     }
 }